Moves NoMatch to component.
