+++ client/src/components/NoMatch/NoMatch.css
@@ -1,3 +1,3 @@
-.error {
+#noMatchError {
   color: red!important;
 }
